+++ gpt4all-chat/src/chatmodel.cpp
@@ -41,6 +41,12 @@ void ChatItem::serializeText(QDataStream &stream, int version)
     stream << value;
 }
 
+void ChatItem::serializeThink(QDataStream &stream, int version)
+{
+    stream << value;
+    stream << thinkingTime;
+}
+
 void ChatItem::serializeSubItems(QDataStream &stream, int version)
 {
     stream << name;
@@ -50,6 +56,7 @@ void ChatItem::serializeSubItems(QDataStream &stream, int version)
         case ToolCall:      { serializeToolCall(stream, version);       break; }
         case ToolResponse:  { serializeToolResponse(stream, version);   break; }
         case Text:          { serializeText(stream, version);           break; }
+        case Think:         { serializeThink(stream, version);          break; }
         case System:
         case Prompt:
             throw std::invalid_argument(fmt::format("cannot serialize subitem type {}", int(typ)));
@@ -162,6 +169,13 @@ bool ChatItem::deserializeResponse(QDataStream &stream, int version)
     return true;
 }
 
+bool ChatItem::deserializeThink(QDataStream &stream, int version)
+{
+    stream >> value;
+    stream >> thinkingTime;
+    return true;
+}
+
 bool ChatItem::deserializeSubItems(QDataStream &stream, int version)
 {
     stream >> name;
@@ -177,6 +191,7 @@ bool ChatItem::deserializeSubItems(QDataStream &stream, int version)
         case ToolCall:      { deserializeToolCall(stream, version); break; }
         case ToolResponse:  { deserializeToolResponse(stream, version); break; }
         case Text:          { deserializeText(stream, version); break; }
+        case Think:         { deserializeThink(stream, version); break; }
         case System:
         case Prompt:
             throw std::invalid_argument(fmt::format("cannot serialize subitem type {}", int(typ)));