+++ gpt4all-chat/src/tool.h
@@ -1,13 +1,16 @@
 #ifndef TOOL_H
 #define TOOL_H
 
+#include <nlohmann/json.hpp>
+
 #include <QList>
 #include <QObject>
 #include <QString>
 #include <QVariant>
 #include <QtGlobal>
 
-#include <jinja2cpp/value.h>
+using json = nlohmann::ordered_json;
+
 
 namespace ToolEnums
 {
@@ -25,6 +28,7 @@ namespace ToolEnums
 
     enum class ParseState {
         None,
+        InTagChoice,
         InStart,
         Partial,
         Complete,
@@ -122,7 +126,7 @@ public:
 
     bool operator==(const Tool &other) const { return function() == other.function(); }
 
-    jinja2::Value jinjaValue() const;
+    json::object_t jinjaValue() const;
 
 Q_SIGNALS:
     void runComplete(const ToolCallInfo &info);